fix points int64=>float64

+++ reward.go
@@ -19,7 +19,7 @@ type discountResult struct {
 
 type RewardService interface {
 	UserRedeemReward(email string, ruleId int64) (*Discount, error)
-	RedeemEarn(email string, ruleType string, points int64) (*EarnedPoints, error)
+	RedeemEarn(email string, ruleType string, points float64) (*EarnedPoints, error)
 	GetUserDiscounts(email string) ([]*Discount, error)
 	GetUserActivities(email string) ([]*UserActivitie, error)
 }
@@ -62,7 +62,7 @@ func (s *RewardServiceOp) UserRedeemReward(email string, ruleId int64) (*Discoun
 }
 
 // 发放积分
-func (s *RewardServiceOp) RedeemEarn(email string, ruleType string, points int64) (*EarnedPoints, error) {
+func (s *RewardServiceOp) RedeemEarn(email string, ruleType string, points float64) (*EarnedPoints, error) {
 	var pointsR *EarnedPoints
 	var errResult *Result
 	_, err := s.client.Client.R().SetResult(&Result{Data: &pointsR}).SetBody(map[string]interface{}{"email": email, "rule_type": ruleType, "points": points}).SetError(&errResult).Post(rewardPath + earnPath)
